backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-dotenv
requests
httpx
pytest
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-dotenv
requests
httpx
pytest
pytest-asyncio
google-generativeai
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
torch
asyncpg
minio
pillow
python-docx
pandas
pymupdf
presidio-analyzer
presidio-anonymizer
transformers
docling-core
platformdirs
typing-extensions
google-generativeai
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
torch
python-dotenv
minio
requests
httpx
pytest
pytest-asyncio

